requirements.txt
python-telegram-bot
requests
redis
Flask
Flask-Cors
psutil
netifaces